+++ http-compiler-plugin/src/main/java/io/ballerina/stdlib/http/compiler/HttpResourceValidator.java
@@ -75,6 +75,7 @@ class HttpResourceValidator {
         extractResourceAnnotationAndValidate(ctx, member);
         extractInputParamTypeAndValidate(ctx, member);
         extractReturnTypeAndValidate(ctx, member);
+        validateHttpCallerUsage(ctx, member);
     }
 
     private static void extractResourceAnnotationAndValidate(SyntaxNodeAnalysisContext ctx,
@@ -447,15 +448,12 @@ class HttpResourceValidator {
         if (returnTypeDescriptorNode.isEmpty()) {
             return;
         }
-        Node returnTypeDescriptor = returnTypeDescriptorNode.get().type();
-        String returnTypeStringValue = returnTypeDescriptor.toString().split(" ")[0];
-
+        String returnTypeStringValue = getReturnTypeDescription(returnTypeDescriptorNode.get());
         Optional<Symbol> functionSymbol = ctx.semanticModel().symbol(member);
         if (functionSymbol.isEmpty()) {
             return;
         }
         FunctionTypeSymbol functionTypeSymbol = ((FunctionSymbol) functionSymbol.get()).typeDescriptor();
-
         Optional<TypeSymbol> returnTypeSymbol = functionTypeSymbol.returnTypeDescriptor();
         returnTypeSymbol.ifPresent(typeSymbol -> validateReturnType(ctx, member, returnTypeStringValue, typeSymbol));
     }
@@ -536,6 +534,61 @@ class HttpResourceValidator {
             return false;
         }
         return expectedType.equals(typeName.get());
+    }
+
+    private static void validateHttpCallerUsage(SyntaxNodeAnalysisContext ctx, FunctionDefinitionNode member) {
+        Optional<Symbol> resourceMethodSymbolOptional = ctx.semanticModel().symbol(member);
+        Optional<ReturnTypeDescriptorNode> returnTypeDescOpt = member.functionSignature().returnTypeDesc();
+        if (resourceMethodSymbolOptional.isEmpty() || returnTypeDescOpt.isEmpty()) {
+            return;
+        }
+        String returnTypeDescription = getReturnTypeDescription(returnTypeDescOpt.get());
+        ResourceMethodSymbol resourceMethod = (ResourceMethodSymbol) resourceMethodSymbolOptional.get();
+        FunctionTypeSymbol typeDescriptor = resourceMethod.typeDescriptor();
+        Optional<List<ParameterSymbol>> parameterOptional = typeDescriptor.params();
+        boolean isCallerPresent = parameterOptional.isPresent() &&
+                parameterOptional.get().stream().anyMatch(HttpResourceValidator::isHttpCaller);
+        if (!isCallerPresent) {
+            return;
+        }
+        Optional<TypeSymbol> returnTypeDescriptorOpt = typeDescriptor.returnTypeDescriptor();
+        if (returnTypeDescriptorOpt.isEmpty()) {
+            return;
+        }
+        TypeSymbol typeSymbol = returnTypeDescriptorOpt.get();
+        if (isValidReturnTypeWithCaller(typeSymbol)) {
+            return;
+        }
+        updateDiagnostic(ctx, member, returnTypeDescription, HttpDiagnosticCodes.HTTP_118);
+    }
+
+    private static boolean isHttpCaller(ParameterSymbol param) {
+        TypeDescKind typeDescKind = param.typeDescriptor().typeKind();
+        if (TypeDescKind.TYPE_REFERENCE.equals(typeDescKind)) {
+            TypeSymbol typeDescriptor = ((TypeReferenceTypeSymbol) param.typeDescriptor()).typeDescriptor();
+            return isHttpModuleType(CALLER_OBJ_NAME, typeDescriptor);
+        }
+        return false;
+    }
+
+    private static String getReturnTypeDescription(ReturnTypeDescriptorNode returnTypeDescriptorNode) {
+        return returnTypeDescriptorNode.type().toString().split(" ")[0];
+    }
+
+    private static boolean isValidReturnTypeWithCaller(TypeSymbol returnTypeDescriptor) {
+        TypeDescKind typeKind = returnTypeDescriptor.typeKind();
+        if (TypeDescKind.UNION.equals(typeKind)) {
+            return ((UnionTypeSymbol) returnTypeDescriptor)
+                    .memberTypeDescriptors().stream()
+                    .map(HttpResourceValidator::isValidReturnTypeWithCaller)
+                    .reduce(true, (a , b) -> a && b);
+        } else if (TypeDescKind.TYPE_REFERENCE.equals(typeKind)) {
+            TypeSymbol typeRef = ((TypeReferenceTypeSymbol) returnTypeDescriptor).typeDescriptor();
+            TypeDescKind typeRefKind = typeRef.typeKind();
+            return TypeDescKind.ERROR.equals(typeRefKind) || TypeDescKind.NIL.equals(typeRefKind);
+        } else {
+            return TypeDescKind.ERROR.equals(typeKind) || TypeDescKind.NIL.equals(typeKind);
+        }
     }
 
     private static void reportInvalidReturnType(SyntaxNodeAnalysisContext ctx, FunctionDefinitionNode node,